Working toward making the number of markers load dynamically. Added a simple data.c3d file (from CMU dataset http://mocap.cs.cmu.edu/subjects/02/02_01.c3d ). Adjusted the map files to the new data. Updated project to be compatible with Qt5. Removed mygraphicsscene (it wasn't doing anything anyway). Fixed (sort of) a bug in the marker widget.
Including a small data.c3d file seems like the thing to do so that
someone can run the project out-of-the-box.  At least that's the goal.

+++ include/quatcamera.h
@@ -56,15 +56,15 @@ public:
   void getTransform(QMatrix4x4& mat);
   void getInvTransform(QMatrix4x4& mat);
 
-  void getRMajorTrans(double* m);
-  void getCMajorTrans(double* m);
-  void getRMajorInvTrans(double* m);
-  void getCMajorInvTrans(double* m);
+  void getRMajorTrans(float* m);
+  void getCMajorTrans(float* m);
+  void getRMajorInvTrans(float* m);
+  void getCMajorInvTrans(float* m);
 
   /// Store quaternion Q in matrix m[16]
   /// If m is a column matrix, this
   /// will store the inverse rotation.
-  void quatToRowMajorMat(const QQuaternion& q,double* m);
+  void quatToRowMajorMat(const QQuaternion& q,float* m);
 
 
   void reset();
